zebrad: use Downloads in sync, and revert to previous sync algorithm.
As designed, the sync algorithm is supposed to proceed by downloading state
prospectively and handle errors by flushing the pipeline and starting over.
This hasn't worked well, because we didn't previously cancel tasks properly.
Now that we can, try to use something in the spirit of the original sync
algorithm.